requirements.txt
# Data Collection
fastf1==3.3.9

# Data Manipulation
pandas==2.2.2
numpy==1.26.4

# Database
mysql-connector-python==8.3.0
SQLAlchemy==2.0.30

# Machine Learning
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0

# Model Explainability
shap==0.45.0

# Experiment Tracking
mlflow==2.13.0

# Visualization
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.22.0

# Dashboard
streamlit==1.35.0

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.1

# Utilities
tqdm==4.66.4
joblib==1.4.2